+++ Runtime/FSM/IFiniteStateMachine.cs
@@ -3,6 +3,7 @@ namespace Slantar.Architecture
 {
 	public interface IFiniteStateMachine
 	{
+		bool ForceState<TState>() where TState : IFSMState;
 		void Update();
 	}
 }