+++ Runtime/FSM/Transitions/AbstractFSMTransition.cs
@@ -1,0 +1,14 @@
+﻿namespace Slantar.Architecture
+{
+    public abstract class AbstractFSMTransition : IFSMTransition
+    {
+        public virtual IFSMState ToState { get; }
+        
+        protected AbstractFSMTransition(IFSMState toState)
+        {
+            ToState = toState;
+        }
+        
+        public abstract bool Valid { get; }
+    }
+}