+++ Runtime/Log/LogFormaters/BasicLogFormatter.cs
@@ -1,0 +1,8 @@
+﻿namespace Slantar.Architecture
+{
+    public class BasicLogFormatter : ILogFormatter
+    {
+        public string Format(LogLevel level, string message) 
+            => $"{level.ToString().ToUpper()}: {message}";
+    }
+}