+++ Runtime/Log/LogFormaters/ILogFormatter.cs
@@ -1,0 +1,7 @@
+﻿namespace Slantar.Architecture
+{
+    public interface ILogFormatter
+    {
+        string Format(LogLevel level, string message);
+    }
+}