+++ Tests/Runtime/FSM/Classes/StateC.cs
@@ -1,0 +1,23 @@
+﻿
+namespace Slantar.Architecture.Tests
+{
+	public class StateC : IFSMState
+	{
+		private Flag boolFlag;
+
+		public StateC(Flag boolFlag) => this.boolFlag = boolFlag;
+
+		public void Enter()
+		{
+			boolFlag.value = false;
+		}
+
+		public void Update()
+		{
+		}
+
+		public void Exit()
+		{
+		}
+	}
+}